requirements.txt
streamlit
streamlit_cropper
Pillow
torchvision
torch
numpy